moved new comment components to the common service
