+++ services/common/src/components/comments/MinistryComment.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from "react";
-import { timeAgo } from "@common/utils/helpers";
+import { timeAgo } from "@mds/core-web/common/utils/helpers";
 
 interface MinistryCommentProps {
   children?: React.ReactNode;
